Move each settings submenu to its own component
